requirements/ci.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements/ci.txt requirements/base.txt requirements/test-tools.in
#
-e git+git://github.com/maykinmedia/djangorestframework-hal.git@f6692d2489645cfd71ebbd59cd2fdb53a311356b#egg=djangorestframework_hal  # via -r requirements/base.txt
beautifulsoup4==4.9.1     # via webtest
certifi==2020.6.20        # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
chardet==3.0.4            # via -r requirements/base.txt, requests
coreapi==2.3.3            # via -r requirements/base.txt, drf-yasg
coreschema==0.0.4         # via -r requirements/base.txt, coreapi, drf-yasg
coverage==4.5.4           # via -r requirements/test-tools.in
cssselect==1.1.0          # via pyquery
django-admin-index==1.4.0  # via -r requirements/base.txt
django-appconf==1.0.4     # via -r requirements/base.txt, django-axes
django-axes==5.4.2        # via -r requirements/base.txt
django-choices==1.7.1     # via -r requirements/base.txt, vng-api-common
django-compat==1.0.15     # via -r requirements/base.txt, django-hijack, django-hijack-admin
django-filter==2.3.0      # via -r requirements/base.txt, vng-api-common
django-hijack-admin==2.1.10  # via -r requirements/base.txt
django-hijack==2.1.10     # via -r requirements/base.txt, django-hijack-admin
django-ipware==3.0.0      # via -r requirements/base.txt, django-axes
django-ordered-model==3.4.1  # via -r requirements/base.txt, django-admin-index
django-redis==4.12.1      # via -r requirements/base.txt
django-rest-framework-condition==0.1.1  # via -r requirements/base.txt, vng-api-common
django-rosetta==0.9.4     # via -r requirements/base.txt
django-sniplates==0.7.0   # via -r requirements/base.txt
django-solo==1.1.3        # via -r requirements/base.txt, vng-api-common
django-webtest==1.9.7     # via -r requirements/test-tools.in
django==2.2.15            # via -r requirements/base.txt, django-appconf, django-axes, django-choices, django-filter, django-redis, django-rest-framework-condition, django-rosetta, django-sniplates, djangorestframework-hal, drf-nested-routers, drf-yasg, vng-api-common
djangorestframework-camel-case==1.2.0  # via -r requirements/base.txt, vng-api-common
djangorestframework==3.9.4  # via -r requirements/base.txt, djangorestframework-hal, drf-nested-routers, drf-yasg, vng-api-common
drf-nested-routers==0.91  # via -r requirements/base.txt, vng-api-common
drf-yasg==1.16.0          # via -r requirements/base.txt, vng-api-common
elastic-apm==5.8.1        # via -r requirements/base.txt
factory-boy==3.0.1        # via -r requirements/base.txt
faker==4.1.3              # via -r requirements/base.txt, factory-boy
freezegun==1.0.0          # via -r requirements/base.txt, -r requirements/test-tools.in
gemma-zds-client==0.13.2  # via -r requirements/base.txt, vng-api-common
idna==2.10                # via -r requirements/base.txt, requests
inflection==0.5.0         # via -r requirements/base.txt, drf-yasg
iso-639==0.4.5            # via -r requirements/base.txt, vng-api-common
isodate==0.6.0            # via -r requirements/base.txt, vng-api-common
itypes==1.2.0             # via -r requirements/base.txt, coreapi
jinja2==2.11.2            # via -r requirements/base.txt, coreschema
lxml==4.5.2               # via pyquery
markupsafe==1.1.1         # via -r requirements/base.txt, jinja2
mock==4.0.2               # via -r requirements/base.txt
oyaml==0.9                # via -r requirements/base.txt, vng-api-common
pep8==1.7.1               # via -r requirements/test-tools.in
pillow==7.2.0             # via -r requirements/base.txt
polib==1.1.0              # via -r requirements/base.txt, django-rosetta
psycopg2==2.8.5           # via -r requirements/base.txt
pyjwt==1.7.1              # via -r requirements/base.txt, gemma-zds-client, vng-api-common
pyquery==1.4.1            # via -r requirements/test-tools.in
python-dateutil==2.8.1    # via -r requirements/base.txt, faker, freezegun
python-decouple==3.3      # via -r requirements/base.txt
python-dotenv==0.14.0     # via -r requirements/base.txt
pytz==2020.1              # via -r requirements/base.txt, django
pyyaml==5.3.1             # via -r requirements/base.txt, gemma-zds-client, oyaml, vng-api-common
redis==3.5.3              # via -r requirements/base.txt, django-redis
requests-mock==1.8.0      # via -r requirements/base.txt, -r requirements/test-tools.in
requests==2.24.0          # via -r requirements/base.txt, coreapi, django-rosetta, gemma-zds-client, requests-mock, vng-api-common
ruamel.yaml.clib==0.2.0   # via -r requirements/base.txt, ruamel.yaml
ruamel.yaml==0.16.10      # via -r requirements/base.txt, drf-yasg
sentry-sdk==0.16.3        # via -r requirements/base.txt
six==1.15.0               # via -r requirements/base.txt, django-choices, django-compat, django-rosetta, drf-yasg, isodate, python-dateutil, requests-mock, webtest
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.3.1           # via -r requirements/base.txt, django
tblib==1.7.0              # via -r requirements/test-tools.in
text-unidecode==1.3       # via -r requirements/base.txt, faker
unidecode==1.1.1          # via -r requirements/base.txt, vng-api-common
uritemplate==3.0.1        # via -r requirements/base.txt, coreapi, drf-yasg
urllib3==1.25.10          # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
uwsgi==2.0.19.1           # via -r requirements/base.txt
vng-api-common==1.5.6     # via -r requirements/base.txt
waitress==1.4.4           # via webtest
webob==1.8.6              # via webtest
webtest==2.0.35           # via django-webtest
xmltodict==0.12.0         # via -r requirements/base.txt
